+++ module-domain/src/main/java/com/yapp/sharefood/food/repository/FoodRepository.java
@@ -2,6 +2,7 @@ package com.yapp.sharefood.food.repository;
 
 import com.yapp.sharefood.category.domain.Category;
 import com.yapp.sharefood.food.domain.Food;
+import com.yapp.sharefood.food.repository.query.FoodQueryRepository;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
