+++ module-domain/src/main/java/com/yapp/sharefood/food/repository/query/FoodTagQueryRepository.java
@@ -1,4 +1,4 @@
-package com.yapp.sharefood.food.repository;
+package com.yapp.sharefood.food.repository.query;
 
 import com.yapp.sharefood.food.domain.FoodTag;
